+++ library/src/main/scala/scopus/se/umu/cs/rai/scopus/collection/package.scala
@@ -1,0 +1,8 @@
+package scopus.se.umu.cs.rai.scopus
+
+/**
+ * Scopus libraries written in Scopus
+ */
+package object collection{
+}
+