+++ library/src/main/scala/se/umu/cs/rai/scopus/collection/package.scala
@@ -1,8 +1,0 @@
-package se.umu.cs.rai.scopus
-
-/**
- * Scopus libraries written in Scopus
- */
-package object collection{
-}
-