+++ translator/src/main/scala/scopus/se/umu/cs/rai/scopus/translator/package.scala
@@ -1,0 +1,7 @@
+package scopus.se.umu.cs.rai.scopus
+
+/**
+ * Translator from Scopus to Scala.
+ */
+package object translator {
+}