+++ translator/src/main/scala/se/umu/cs/rai/scopus/translator/package.scala
@@ -1,7 +1,0 @@
-package se.umu.cs.rai.scopus
-
-/**
- * Translator from Scopus to Scala.
- */
-package object translator {
-}