+++ translator/src/test/resources/example/Example.scala
@@ -1,5 +1,0 @@
-package example
-
-case class Example() {
-
-}