+++ translator/src/test/resources/example/otherexample/OtherExample.scala
@@ -1,5 +1,0 @@
-package example.otherexample
-
-case class OtherExample() {
-
-}