+++ translator/src/test/resources/scopus/example/otherexample/OtherExample.scala
@@ -1,0 +1,5 @@
+package scopus.example.otherexample
+
+case class OtherExample() {
+
+}