+++ translator/src/test/resources/scopus/se/umu/cs/rai/scopus/translator/example/HelloWorld.scala
@@ -1,0 +1,13 @@
+package scopus.se.umu.cs.rai.scopus.translator.example
+
+case class Main() {
+
+  def main (args: Array[String]) =
+    println("Hello world!")
+
+}
+
+object EntryPoint {
+  def main(args: Array[String]): Unit = Main().main(args)
+}
+