+++ translator/src/test/scala/scopus/example/Example.scala
@@ -1,0 +1,5 @@
+package scopus.example
+
+case class Example() {
+
+}