+++ translator/src/test/scala/se/umu/cs/rai/scopus/translator/example/SwapExample.scala
@@ -1,8 +1,0 @@
-package se.umu.cs.rai.scopus.translator.example
-
-case class My_example() {
-
-  def swap (pair: (Int, Int)) =
-    (pair._2, pair._1)
-
-}